+++ app/src/main/res/layout/activity_planet_detail.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.view.settings.PlanetDetailActivity">
+    tools:context=".presentation.view.main.planet_list.PlanetDetailActivity">
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
