+++ app/src/main/res/layout/fragment_daily_record.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.view.settings.DailyRecordFragment">
+    tools:context=".presentation.view.main.daily_record.DailyRecordFragment">
 
     <TextView
         android:layout_width="wrap_content"
