+++ app/src/main/res/layout/fragment_my_page.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.view.settings.MyPageFragment">
+    tools:context=".presentation.view.main.my_page.MyPageFragment">
 
     <TextView
         android:layout_width="wrap_content"
