+++ app/src/main/res/layout/fragment_planet_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.view.settings.PlanetListFragment">
+    tools:context=".presentation.view.main.planet_list.PlanetListFragment">
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
