+++ app/src/main/res/navigation/bottom_nav_graph.xml
@@ -7,22 +7,22 @@
 
     <fragment
         android:id="@+id/action_home"
-        android:name="com.sesac.planet.presentation.view.settings.HomeFragment"
+        android:name="com.sesac.planet.presentation.view.main.home.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/action_planet_list"
-        android:name="com.sesac.planet.presentation.view.settings.PlanetListFragment"
+        android:name="com.sesac.planet.presentation.view.main.planet_list.PlanetListFragment"
         android:label="fragment_planet_list"
         tools:layout="@layout/fragment_planet_list" />
     <fragment
         android:id="@+id/action_daily_record"
-        android:name="com.sesac.planet.presentation.view.settings.DailyRecordFragment"
+        android:name="com.sesac.planet.presentation.view.main.daily_record.DailyRecordFragment"
         android:label="fragment_daily_record"
         tools:layout="@layout/fragment_daily_record" />
     <fragment
         android:id="@+id/action_my_page"
-        android:name="com.sesac.planet.presentation.view.settings.MyPageFragment"
+        android:name="com.sesac.planet.presentation.view.main.my_page.MyPageFragment"
         android:label="fragment_my_page"
         tools:layout="@layout/fragment_my_page" />
 </navigation>
